+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 
 - Improved Read me
 
+- Improved template for new projects
+
 - "Apply all code actions" command renamed to "Apply all code actions in Umajin project" to be searchable by "umajin" keyword
 
 ## [1.1.4] - 2022-06-30